Renamed Plugin to BasePlugin to accurately reflect what it is. Added default methods.
